Remove Vector from default LocalSettings
Which skin is used should be up to the user. And manage in their
LocalSettings.php after including the .docker/LocalSettings.php file.

That way the default page after cloning MediaWiki will be the
fallback skin telling the user to add wfLoadSkin(), instead of
a PHP-level error saying mmediawiki/skins/Vector doesn't exist.